refactor: extract index.ts concerns into dedicated modules
