+++ backend/src/index.ts
@@ -1,11 +1,12 @@
 // src/index.ts
-import { Elysia, t } from "elysia";
+import { Elysia } from "elysia";
 import { cors } from "@elysiajs/cors";
 import { swagger } from "@elysiajs/swagger";
 import { config } from "./config";
 import { db } from "./db";
 import { authMiddleware } from "./middleware/auth";
 import { rateLimiters } from "./middleware/rate-limit";
+import { requestLimitsMiddleware } from "./middleware/request-limits";
 import {
   correlationMiddleware,
   enhancedApiLogging,
@@ -22,22 +23,14 @@ import { goalRoutes } from "./modules/goals/routes";
 import { habitRoutes } from "./modules/habits/routes";
 import { billingRoutes } from "./modules/billing/routes";
 import { reportingRoutes } from "./modules/reporting/routes";
+import { webhookHandler } from "./modules/billing/webhook-handler";
+import { healthRoutes } from "./routes/health";
 
 logger.info("🚀 Starting Elysia server...");
 
 const app = new Elysia()
   // Request size limits for security
-  .onRequest(({ request, set }) => {
-    const contentLength = request.headers.get("content-length");
-    if (contentLength && parseInt(contentLength) > 1024 * 1024) {
-      // 1MB limit
-      set.status = 413;
-      return {
-        code: "PAYLOAD_TOO_LARGE",
-        message: "Request body too large. Maximum size is 1MB.",
-      };
-    }
-  })
+  .use(requestLimitsMiddleware)
 
   // Global middleware & plugins
   .use(
@@ -75,240 +68,7 @@ if (config.NODE_ENV !== "production") {
 app.decorate("db", db);
 
 // Webhook routes (NO AUTH) - MUST be before middleware that consumes body
-app.post(
-  "/api/billing/webhook",
-  async (ctx: any) => {
-    const { logger } = await import("./lib/logger");
-    try {
-      // Get the raw body text from custom parse function
-      const rawBodyText = ctx.body;
-      const signature = ctx.headers["stripe-signature"];
-
-      logger.info(
-        {
-          operation: "stripe_webhook",
-          hasRawBody: !!rawBodyText,
-          rawBodyLength: rawBodyText?.length || 0,
-          hasSignature: !!signature,
-          signature: signature ? signature.substring(0, 20) + "..." : "none",
-          rawBodyPreview:
-            rawBodyText ? rawBodyText.substring(0, 100) + "..." : "none",
-        },
-        "Received Stripe webhook request"
-      );
-
-      if (!signature) {
-        logger.error(
-          { operation: "stripe_webhook", error: "Missing Stripe signature" },
-          "Missing Stripe signature header"
-        );
-        ctx.set.status = 400;
-        return { received: false, error: "Missing Stripe signature" };
-      }
-
-      if (!rawBodyText || rawBodyText.length === 0) {
-        logger.error(
-          {
-            operation: "stripe_webhook",
-            error: "Missing or empty request body",
-          },
-          "Missing or empty request body for webhook"
-        );
-        ctx.set.status = 400;
-        return { received: false, error: "Missing request body" };
-      }
-
-      // Import services inline to avoid circular dependencies
-      const { StripeService } = await import(
-        "./modules/billing/stripe-service"
-      );
-      const { SubscriptionService } = await import(
-        "./modules/billing/subscription-service"
-      );
-
-      // Verify webhook signature and normalize event format
-      let normalizedEvent;
-      try {
-        normalizedEvent = await StripeService.verifyWebhookSignature(
-          rawBodyText,
-          signature
-        );
-      } catch (err) {
-        logger.error(
-          { operation: "stripe_webhook", error: err },
-          "Signature verification or event parsing failed"
-        );
-        ctx.set.status = 400;
-        return { received: false, error: "Signature verification failed" };
-      }
-
-      // Deduplication: check if event ID already processed
-      const eventId = normalizedEvent.id;
-      const eventExists = ctx.db
-        .prepare("SELECT 1 FROM stripe_events WHERE id = ?")
-        .get(eventId);
-      if (eventExists) {
-        logger.warn(
-          { operation: "stripe_webhook", eventId },
-          `Duplicate Stripe event received, skipping: ${eventId}`
-        );
-        ctx.set.status = 200;
-        return { received: true, duplicate: true, eventId };
-      }
-
-      logger.info(
-        {
-          operation: "stripe_webhook",
-          eventType: normalizedEvent.type,
-          eventId: normalizedEvent.id,
-          format: normalizedEvent.format,
-          payload: normalizedEvent,
-        },
-        `Processing Stripe webhook: ${normalizedEvent.type} (${normalizedEvent.format})`
-      );
-
-      // Handle only supported event types
-      if (
-        normalizedEvent.type &&
-        normalizedEvent.type.includes("subscription")
-      ) {
-        let subscription;
-        if (
-          normalizedEvent.format === "thin" &&
-          normalizedEvent.related_object
-        ) {
-          subscription = await StripeService.fetchRelatedObject(
-            normalizedEvent.related_object
-          );
-        } else if (normalizedEvent.data?.object) {
-          subscription = normalizedEvent.data.object;
-        }
-
-        if (subscription) {
-          const customerId = subscription.customer;
-          const subscriptionId = subscription.id;
-          const status = subscription.status;
-
-          // Find user by Stripe customer ID
-          const user = ctx.db
-            .prepare("SELECT id FROM users WHERE stripe_customer_id = ?")
-            .get(customerId) as { id: number } | undefined;
-
-          if (user) {
-            if (
-              status === "canceled" ||
-              normalizedEvent.type.includes("deleted")
-            ) {
-              await SubscriptionService.cancelSubscription(
-                user.id,
-                subscriptionId
-              );
-              logger.info(
-                {
-                  operation: "stripe_webhook_processed",
-                  eventType: normalizedEvent.type,
-                  userId: user.id,
-                  subscriptionId,
-                  eventId,
-                },
-                "Canceled subscription from webhook"
-              );
-            } else {
-              const subscriptionItem = subscription.items.data[0];
-              if (!subscriptionItem) {
-                throw new Error("Subscription has no items");
-              }
-              const currentPeriodEnd = new Date(
-                subscriptionItem.current_period_end * 1000
-              ).toISOString();
-              await SubscriptionService.upsertSubscription(
-                user.id,
-                subscriptionId,
-                status,
-                currentPeriodEnd
-              );
-              logger.info(
-                {
-                  operation: "stripe_webhook_processed",
-                  eventType: normalizedEvent.type,
-                  userId: user.id,
-                  subscriptionId,
-                  status,
-                  eventId,
-                },
-                "Updated subscription from webhook"
-              );
-            }
-          } else {
-            logger.warn(
-              {
-                operation: "stripe_webhook",
-                eventType: normalizedEvent.type,
-                customerId,
-                subscriptionId,
-                eventId,
-              },
-              "User not found for Stripe customer ID"
-            );
-          }
-        }
-      } else {
-        // Log and gracefully handle unsupported event types
-        logger.info(
-          {
-            operation: "stripe_webhook",
-            eventType: normalizedEvent?.type,
-            eventId,
-          },
-          `Received unsupported or unhandled event type: ${normalizedEvent?.type}`
-        );
-      }
-
-      // Insert event ID into stripe_events for deduplication
-      ctx.db.prepare("INSERT INTO stripe_events (id) VALUES (?)").run(eventId);
-
-      ctx.set.status = 200;
-      return {
-        received: true,
-        format: normalizedEvent.format,
-        eventType: normalizedEvent.type,
-        eventId,
-      };
-    } catch (error) {
-      logger.error(
-        {
-          error: error instanceof Error ? error : new Error(String(error)),
-          operation: "stripe_webhook",
-        },
-        "Failed to process webhook"
-      );
-      ctx.set.status = 200;
-      return { received: false, error: "Webhook processing failed" };
-    }
-  },
-  {
-    // Custom parse function to handle Stripe webhook raw body
-    async parse(ctx) {
-      // Check for Stripe webhook content-type with charset
-      const contentType = ctx.request.headers.get("content-type");
-      if (contentType === "application/json; charset=utf-8") {
-        const reqText = await ctx.request.text();
-        return reqText;
-      } else {
-        // For other content types, let Elysia handle normally
-        return undefined;
-      }
-    },
-    headers: t.Object({
-      "stripe-signature": t.String(),
-    }),
-    body: t.String(),
-    detail: {
-      summary: "Handle Stripe webhooks (NO AUTH)",
-      tags: ["Billing"],
-    },
-  }
-);
+app.use(webhookHandler);
 
 // Apply middleware after webhook routes to avoid body consumption conflicts
 app.use(correlationMiddleware);
@@ -331,87 +91,8 @@ app.use(habitRoutes);
 app.use(billingRoutes);
 app.use(reportingRoutes);
 
-// Root endpoint
-app.get(
-  "/",
-  () => ({
-    status: "ok",
-    message: "Macro Trackr API is running!",
-    timestamp: new Date().toISOString(),
-  }),
-  {
-    detail: { summary: "API Root / Health Check", tags: ["System"] },
-  }
-);
-
-// Health check endpoint for monitoring
-app.get(
-  "/health",
-  () => {
-    try {
-      // Test database connectivity
-      const dbCheck = db.prepare("SELECT 1 as health").get() as
-        | { health: number }
-        | undefined;
-
-      return {
-        status: "healthy",
-        timestamp: new Date().toISOString(),
-        version: "1.0.0",
-        environment: config.NODE_ENV,
-        database: dbCheck?.health === 1 ? "connected" : "disconnected",
-      };
-    } catch (error) {
-      logger.error({ error }, "Health check failed");
-      return {
-        status: "unhealthy",
-        timestamp: new Date().toISOString(),
-        version: "1.0.0",
-        environment: config.NODE_ENV,
-        database: "error",
-        error: error instanceof Error ? error.message : "Unknown error",
-      };
-    }
-  },
-  {
-    detail: {
-      summary: "Health Check Endpoint for Load Balancers",
-      tags: ["System"],
-    },
-  }
-);
-
-// Readiness probe for Kubernetes
-app.get(
-  "/health/ready",
-  () => {
-    try {
-      // Check if all dependencies are ready
-      const dbCheck = db.prepare("SELECT 1 as ready").get() as
-        | { ready: number }
-        | undefined;
-
-      if (dbCheck?.ready === 1) {
-        return { status: "ready", timestamp: new Date().toISOString() };
-      } else {
-        return { status: "not ready", reason: "database not ready" };
-      }
-    } catch (error) {
-      logger.error({ error }, "Readiness check failed");
-      return {
-        status: "not ready",
-        reason: "database error",
-        error: error instanceof Error ? error.message : "Unknown error",
-      };
-    }
-  },
-  {
-    detail: {
-      summary: "Readiness Probe for Container Orchestration",
-      tags: ["System"],
-    },
-  }
-);
+// Health check routes (public, no auth)
+app.use(healthRoutes);
 
 // Global error handling
 app.onError(({ code, error, set, path }: any) => {
